Move `Lrama::Report::Profile` to `Lrama::Report::Profile::CallStack`
