+++ lib/lrama/command.rb
@@ -7,7 +7,7 @@ module Lrama
 
     def run(argv)
       parser_options(argv)
-      Lrama::Report::Profile.report_profile(@options.profile_opts[:call_stack]) do
+      Lrama::Report::Profile::CallStack.report_profile(@options.profile_opts[:call_stack]) do
         _run
       end
     end
